[UPD] re-organize components
