+++ src/components/reporter/findings/ContestWarning.js
