+++ src/components/reporter/findings/fields.js
