+++ src/components/reporter/findings/functions.js
