+++ src/components/reporter/findings/state.js
