+++ src/components/reporter/widgets/SelectField.js
@@ -1,7 +1,7 @@
 import React, { useCallback } from "react";
 import clsx from "clsx";
 import Select from "react-select";
-
+import ContestWarning from "../findings/ContestWarning";
 import * as formStyles from "./WardenField.module.scss";
 
 const SelectFieldOptionLabel = ({ label }) => {
@@ -29,16 +29,24 @@ const SelectField = ({
   );
 
   return (
-    <Select
-      name={name}
-      required={required}
-      value={options.find((o) => o.value === fieldState) || "Select ..."}
-      formatOptionLabel={SelectFieldOptionLabel}
-      options={options}
-      onChange={handleChange}
-      className={clsx(formStyles.ReactSelect, isInvalid && formStyles.Invalid)}
-      classNamePrefix="react-select"
-    />
+    <>
+      <Select
+        name={name}
+        required={required}
+        value={options.find((o) => o.value === fieldState) || "Select ..."}
+        formatOptionLabel={SelectFieldOptionLabel}
+        options={options}
+        onChange={handleChange}
+        className={clsx(
+          formStyles.ReactSelect,
+          isInvalid && formStyles.Invalid
+        )}
+        classNamePrefix="react-select"
+      />
+      {(fieldState === "G (Gas Optimization)" ||
+        fieldState === "QA (Quality Assurance)") &&
+        name === "risk" && <ContestWarning />}
+    </>
   );
 };
